requirements (3).txt
# ---- Core framework ----
streamlit
python-dotenv

# ---- LangChain core ----
langchain
langchain-core
langchain-community
langchain-text-splitters

# ---- PDF and document parsing ----
pypdf
PyMuPDF
pdfplumber

# ---- Optional: text splitting / utilities ----
tiktoken

# ---- Compatibility helpers ----
requests

langchain-google-genai